+++ packages/carbon-graphs/src/js/index.js
@@ -1,0 +1,3 @@
+import Carbon from './carbon';
+
+export default Carbon;