+++ packages/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDataColorWeightBased.js
@@ -1,5 +1,5 @@
 import { loadBubblePopup } from '@cerner/terra-graphs-docs/lib/example-datasets/popup.js';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import Carbon from '@cerner/carbon-graphs';
 
 const bubbleDataColorWeightBased = {
   key: 'bubble_color_weight_dataset',