+++ packages/terra-graphs-docs/src/example-datasets/dataObjects/Gantt/dynamicPanningDataset1.js
@@ -1,5 +1,5 @@
 import { loadTaskPopup, loadXAndYAxisLabelPopup } from '@cerner/terra-graphs-docs/lib/example-datasets/popup';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import Carbon from '@cerner/carbon-graphs';
 import { loadPopup } from '../../popup';
 
 const scheduled = {